bones_bevy_renderer

Module render

Contains the systems for extracting, syncing, and loading bones renderables.

Structs

BevyBonesEntity

Marker component for entities that are rendered in Bevy for bones.

BonesGameEntity

Resource containing the entity spawned for all of the bones game renderables.

BonesImageIds

Resource mapping bones image IDs to their bevy handles.

Functions

extract_bones_sprites

Extracts the [bones::Sprite]s to be used in bevy’s rendering.

extract_bones_tilemaps

Extracts the [bones::TileLayer]s & [bones::Tile]s to be used in bevy’s rendering.

sync_bones_path2ds

Syncs the [bones::Path2d]s with bevy’s [lyon::Path]s.

sync_bones_window

Syncs elements of the bones window

sync_cameras

Sync bones cameras with Bevy

sync_clear_color

Updates the BonesGame’s [bones::ClearColor] with the [ClearColor] from bevy.