Trait bones_bevy_renderer::HasBonesWorld

pub trait HasBonesWorld: Resource {
    fn world(&mut self) -> &mut World;
}

This is a trait that must be implemented for your Bevy resource containing the bones [World][bones::World].

It gives the BonesRendererPlugin a way to know how to read the bones world from your world resource.

Required Methods

fn world(&mut self) -> &mut World

Return a mutable reference to the bones world stored by the resource.